refactor: move run-fix endpoint to ExpeditionListController at /api/expedition-list/run-fix
